Migrate java9 tests to their own class, run java 8 on java 8
