+++ user/test/com/google/gwt/dev/jjs/CompilerSuite.java
@@ -31,6 +31,7 @@ import com.google.gwt.dev.jjs.test.Java10Test;
 import com.google.gwt.dev.jjs.test.Java11Test;
 import com.google.gwt.dev.jjs.test.Java7Test;
 import com.google.gwt.dev.jjs.test.Java8Test;
+import com.google.gwt.dev.jjs.test.Java9Test;
 import com.google.gwt.dev.jjs.test.JavaAccessFromJavaScriptTest;
 import com.google.gwt.dev.jjs.test.JsniConstructorTest;
 import com.google.gwt.dev.jjs.test.JsniDispatchTest;
@@ -73,6 +74,7 @@ public class CompilerSuite {
     // Java8Test cannot be the first one in a suite. It uses a hack
     // to avoid executing if not in a Java 8+ environment.
     suite.addTestSuite(Java8Test.class);
+    suite.addTestSuite(Java9Test.class);
     suite.addTestSuite(Java10Test.class);
     suite.addTestSuite(Java11Test.class);
     suite.addTestSuite(JavaAccessFromJavaScriptTest.class);
